+++ packages/database/prisma/migrations/20260828171008_m9_c1_voucher_kategori/migration.sql
@@ -1,0 +1,8 @@
+-- AlterTable
+ALTER TABLE "PromoCode" ADD COLUMN     "categoryId" TEXT;
+
+-- CreateIndex
+CREATE INDEX "PromoCode_categoryId_idx" ON "PromoCode"("categoryId");
+
+-- AddForeignKey
+ALTER TABLE "PromoCode" ADD CONSTRAINT "PromoCode_categoryId_fkey" FOREIGN KEY ("categoryId") REFERENCES "Category"("id") ON DELETE SET NULL ON UPDATE CASCADE;